+++ docs/CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ### Patch Changes
 
+- Updated dependencies [be410bd]
+- Updated dependencies [36fd8c7]
+- Updated dependencies [64cec8f]
+  - nextra-theme-docs@2.2.7
+  - nextra@2.2.7
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [be24334]
 - Updated dependencies [4fc13df]
   - nextra@2.2.6